requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
identify==2.6.3
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.23
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.0
openai==1.57.1
orjson==3.10.12
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
propcache==0.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
sniffio==1.3.1
soupsieve==2.6
spotipy==2.24.0
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.1.3
yarl==1.18.3
